+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 * Remove "save screen" option (always on)
 * Change again the threshold for triggering a non-draw frame
+* Don't scroll input editor when rewinding (#381)
 
 ### Fixed